Base-Drift-Code/requirements.txt
torch>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
python-Levenshtein>=0.21.0